requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --output-file requirements.txt
jinja2==3.1.6
    # via -r requirements.in
lightningcss==0.3.1
    # via -r requirements.in
markdown==3.10
    # via -r requirements.in
markupsafe==3.0.3
    # via jinja2
pillow==12.0.0
    # via -r requirements.in
pyyaml==6.0.3
    # via -r requirements.in
tqdm==4.67.1
    # via -r requirements.in
